Server/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorlog==4.1.0
elasticsearch==7.5.1
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python3-logstash==0.4.80
random-user-agent==1.0.1
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
